Adding and renaming
I added 2 of my own designs: a sticker of the "100% unhackable" meme with a lock icon, and a simple version of "IOT: the S stands for security." The IOT one also has a version where the text is outlined in white, so it can be put onto any background. All images are pngs, so they are easily editable.